+++ frontend/src/features/chat/components/ChatHistory.tsx
@@ -1,5 +1,3 @@
-import { useLayoutEffect } from 'react';
-
 import { Container } from '@/components/Layout/Container';
 import { ContainerItem } from '@/components/Layout/ContainerItem';
 
@@ -7,19 +5,11 @@ import { handleMessageDto } from '../types/MessageDto';
 
 export const ChatHistory = ({
   chatHistMsgs,
-  isNeedScroll,
   scrollBottomRef,
 }: {
   chatHistMsgs: handleMessageDto[];
-  isNeedScroll: boolean;
   scrollBottomRef: React.RefObject<HTMLDivElement>;
 }) => {
-  useLayoutEffect(() => {
-    if (isNeedScroll === true) {
-      scrollBottomRef?.current?.scrollIntoView();
-    }
-  }, [chatHistMsgs]);
-
   return (
     <Container flexDirection={'column'}>
       <ContainerItem overflowY={'scroll'}>
