+++ frontend/src/hooks/useScroll.ts
@@ -1,0 +1,50 @@
+import { useState, useLayoutEffect, useRef } from 'react';
+
+const isScrollBottom = (scrollBottomRef: React.RefObject<HTMLDivElement>) => {
+  const scrollParentElement = scrollBottomRef?.current?.parentElement;
+
+  // TODO : is it ok to return false?
+  if (scrollParentElement === undefined || scrollParentElement === null) {
+    return false;
+  }
+  const scrollHeight = scrollParentElement.scrollHeight;
+  const scrollTop = scrollParentElement.scrollTop;
+  const clientHeight = scrollParentElement.clientHeight;
+
+  // debug
+  // console.log("scrollHeight", scrollHeight);
+  // console.log("scrollTop", scrollTop);
+  // console.log("clientHeight", clientHeight);
+  // console.log(scrollTop + clientHeight);
+
+  // TODO : is it ok to return false?
+  if (
+    scrollHeight === undefined ||
+    scrollTop === undefined ||
+    clientHeight === undefined
+  ) {
+    return false;
+  }
+
+  const isBtm = Math.abs(scrollHeight - (scrollTop + clientHeight)) <= 1;
+
+  return isBtm;
+};
+
+export const useScroll = (deps?: React.DependencyList | undefined) => {
+  const scrollBottomRef = useRef<HTMLDivElement>(null);
+  const [isNeedScroll, setIsNeedScroll] = useState(false);
+
+  const handleScroll = () => {
+    const needScroll = isScrollBottom(scrollBottomRef);
+    setIsNeedScroll(() => needScroll);
+  };
+
+  useLayoutEffect(() => {
+    if (isNeedScroll === true) {
+      scrollBottomRef?.current?.scrollIntoView();
+    }
+  }, [isNeedScroll, scrollBottomRef, deps]);
+
+  return { scrollBottomRef, handleScroll };
+};
